+++ packages/snap-preact-components/tests/cypress/component/molecules/recommendation.cy.tsx
@@ -72,7 +72,8 @@ describe('Recommendation Component', async () => {
 		);
 		cy.get('.ss__recommendation').should('exist');
 		cy.get('.ss__recommendation .findMe .result').should('have.length', controller.store.results.length);
-		cy.get('@render').should('have.been.calledOnce');
+		// should be called 20 times (one for each result rendered)
+		cy.get('@render').its('callCount').should('eq', 20);
 		cy.wait(3000);
 		cy.get('@impression').its('callCount').should('eq', 5);
 	});